+++ ui/src/main/java/io/snabble/sdk/ui/checkout/CheckoutActivity.kt
@@ -8,6 +8,7 @@ import android.os.Bundle
 import android.view.View
 import android.view.ViewGroup
 import android.view.WindowManager
+import android.widget.FrameLayout
 import androidx.activity.addCallback
 import androidx.appcompat.app.AlertDialog
 import androidx.core.view.ViewCompat
@@ -23,12 +24,15 @@ import androidx.navigation.NavController
 import androidx.navigation.NavGraph
 import androidx.navigation.NavOptions
 import androidx.navigation.fragment.NavHostFragment
+import com.google.android.material.appbar.MaterialToolbar
 import io.snabble.sdk.InitializationState
 import io.snabble.sdk.PaymentMethod
 import io.snabble.sdk.Snabble
 import io.snabble.sdk.checkout.Checkout
 import io.snabble.sdk.checkout.CheckoutState
 import io.snabble.sdk.ui.R
+import io.snabble.sdk.ui.remotetheme.onToolBarColorForProject
+import io.snabble.sdk.ui.remotetheme.toolBarColorForProject
 import io.snabble.sdk.utils.Logger
 
 class CheckoutActivity : FragmentActivity() {
@@ -136,9 +140,24 @@ class CheckoutActivity : FragmentActivity() {
     private fun setUpToolBarAndStatusBar() {
         val showToolBar = resources.getBoolean(R.bool.showToolbarInCheckout)
         findViewById<View>(R.id.checkout_toolbar_spacer)?.isVisible = showToolBar
+
         navController.addOnDestinationChangedListener { _, _, arguments ->
-            findViewById<View>(R.id.checkout_toolbar)?.isVisible = arguments?.getBoolean("showToolbar", false) == true
+            findViewById<View>(R.id.checkout_toolbar)?.isVisible =
+                arguments?.getBoolean("showToolbar", false) == true
+
+            val toolBarColor =
+                this@CheckoutActivity.toolBarColorForProject(Snabble.checkedInProject.value)
+            val onToolBarColor =
+                this@CheckoutActivity.onToolBarColorForProject(Snabble.checkedInProject.value)
+            this.findViewById<MaterialToolbar>(R.id.checkout_toolbar).apply {
+                toolBarColor?.let(::setBackgroundColor)
+                onToolBarColor?.let(::setTitleTextColor)
+            }
+            this.findViewById<FrameLayout>(R.id.checkout_toolbar_spacer).apply {
+                toolBarColor?.let(::setBackgroundColor)
+            }
         }
+
         if (showToolBar) {
             applyInsets()
         }